update fmt(1) wrappers to work with 9'd PATH
plan9port includes Plan 9's fmt(1), which is incompatible with these
scripts.  Normally that would not be an issue, since $PLAN9/bin is
supposed to be at the end of the PATH.  However, the 9 script puts
$PLAN9/bin at the front of the PATH.  If plan9port's sam editor is
started with "9 sam" (to distinguish it from deadpixi/sam), then
Plan 9's fmt(1) will be at the front of the path, and using these
scripts within sam would fail.  As a workaround, detect that
situation, and use the full path to /usr/bin/fmt if necessary.