Add fmts: fmt(1) "smart" wrapper script
fmts examines the input text.  If it starts with a bullet point token,
the text is wrapped with fmtbl.  If it starts with a line comment or
Markdown quotation token, the text is wrapped with fmtp.  Otherwise,
the text is wrapped with fmtw.

Update 72 and 80 to be fmts aliases rather than fmtw aliases.

Add -p option to fmtp so that fmtp doesn't need to re-detect the prefix
when called from fmts.

Update README.md to:

- Document fmts
- Group alias scripts like 80 with the script being aliased
- Re-order the list: abandon alphabetization and instead put related
  scripts next to each other in the list