Add and use setpipefail script
Move the conditional "set -o pipefail" into a script that other scripts
can source.  This will make it easier to update later, if dash adds
support for pipefail or if this repository needs to be updated to work
with some other shell which doesn't support it.

All the old if conditions were wrong, anyway.  It was supposed to be
"if !linux || !dash", not "linux && !dash", since pipefail is supported
by both the FreeBSD sh and the OpenBSD sh.